modify anonymous manager class
