+++ src/utils/criteriaCompletionChecker.ts
@@ -1,0 +1,6 @@
+export class CriteriaCompletionChecker {
+  getMatchedCriteria(criteriaData: string, localStoredEventList: string) {
+    console.error('criteriaData', criteriaData);
+    console.error('localStoredEventList', localStoredEventList);
+  }
+}
